pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "work-data-hub-pro"
version = "0.1.0"
description = "Capability-first rebuild for WorkDataHub"
requires-python = ">=3.12"
dependencies = [
  "openpyxl>=3.1,<4.0",
  "typer>=0.12,<1.0",
]

[dependency-groups]
dev = [
  "PyYAML>=6.0,<7.0",
  "pytest>=8.2,<9.0",
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "--basetemp=.pytest_tmp"
